+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Add boot support for Ventoy.
 - Change GRUB boot loader to use labels.
 - Change Bash scripts to be verified by ShellCheck.
+- Change Firefox ESR and Google Chrome to use libeatmydata for faster speeds and lower writes.
 
 ## 4.1.0 - The Dual-Boot Update